move container & section to LayoutComponets
