+++ src/components/Hero/Hero.jsx
@@ -1,7 +1,5 @@
 import Image from "next/image";
-// import Buttons from "../Buttons/Buttons";
-import Container from "@/components/Container/Container";
-import Section from "../Section/Section";
+import { Container, Section } from "../LayoutComponents";
 
 export default function Hero() {
   return (
