+++ src/components/LayoutComponents/index.js
@@ -1,0 +1,2 @@
+export { default as Container } from "./Container/Container.jsx";
+export { default as Section } from "./Section/Section.jsx";
